+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- **[UPDATE]** Add `background` prop to `CheckShieldIcon`
 - [...]
 
 # v58.0.0 (25/03/2021)